# openapi.yml
openapi: 3.0.3
info:
  title: buff
  description: ""
  license:
    name: ""
  version: 0.1.0
paths:
...
  /api/admin/order/all-in-period:
    get:
      tags:
        - handlers::admin::orders
      operationId: all_in_period
      parameters:
        - name: start_datetime
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_datetime
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Orders were successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
  /api/admin/order/{id}/cancel:
    patch:
      tags:
        - handlers::admin::orders
      operationId: cancel_order_by_id
      parameters:
        - name: id
          in: path
          description: Order id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Order was successfully canceled
        "400":
          description: Order has already been marked as succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Order was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
  /api/admin/order/{id}/success:
    patch:
      tags:
        - handlers::admin::orders
      operationId: finish_order_by_id
      parameters:
        - name: id
          in: path
          description: Order id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Order was marked as succeeded
        "400":
          description: Order has already been marked as canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Order was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
  /api/admin/review/video:
    post:
      tags:
        - handlers::admin::reviews
      operationId: add_video_review
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddVideoReviewRequest"
        required: true
      responses:
        "204":
          description: Video review was successfully saved
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "409":
          description: Provided url has already been added to video reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
    delete:
      tags:
        - handlers::admin::reviews
      operationId: remove_video_review
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveVideoReviewRequest"
        required: true
      responses:
        "204":
          description: Video review was successfully deleted
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Provided review id does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
  /api/auth/admin/login:
    post:
      tags:
        - handlers::auth::admins
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
        required: true
      responses:
        "200":
          description: Admin or moderator was successfully authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Bad username or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
...
components:
  schemas:
...
    AddVideoReviewRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
    AdminLoginResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
...
    Credentials:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
        password:
          type: string
...
    Details:
      type: object
      required:
        - details
      properties:
        details:
          type: string
...
    Order:
      type: object
      required:
        - id
        - payment_method
        - status
        - created_at
        - steam_id
        - amount
        - fixed_currency_rate
        - currency_symbol
      properties:
        id:
          type: integer
          format: int64
        payment_method:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        steam_id:
          type: integer
          format: int64
        moderator_id:
          type: integer
          format: int64
          nullable: true
        amount:
          type: string
        fixed_currency_rate:
          type: string
        currency_symbol:
          type: string
        finished_at:
          type: string
          format: date-time
          nullable: true
    RemoveVideoReviewRequest:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
...
  securitySchemes:
    jwt_admin:
      type: apiKey
      in: header
      name: X-AM-Authorization
    jwt_user:
      type: http
      scheme: bearer
      bearerFormat: JWT